+++ src/ic-cdk/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- Add `reserved_cycles` and `reserved_cycles_limit` to the management canister.
+
 ## [0.12.0] - 2023-11-23
 
 ### Changed